feat: 자유게시판 페이지 SSR로 변경
